backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "openai>=1.108.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
    "uvicorn[standard]>=0.36.0",
    "waitress>=3.0.2",
]
